+++ CrystalAlgo/CrystalPhysics/PBSPHBoundarySolver.cpp
@@ -115,10 +115,10 @@ float PBSPHBoundarySolver::getOverX(const float x)
 {
 	float over = 0;
 	if (x > boundary.getMaxX()) {
-		over = x - boundary.getMaxX();
+		over = x - static_cast<float>(boundary.getMaxX());
 	}
 	else if (x < boundary.getMinX()) {
-		over = x - boundary.getMinX();
+		over = x - static_cast<float>(boundary.getMinX());
 	}
 	return over;
 }
@@ -127,10 +127,10 @@ float PBSPHBoundarySolver::getOverY(const float y)
 {
 	float over = 0;
 	if (y > boundary.getMaxY()) {
-		over = y - boundary.getMaxY();
+		over = y - static_cast<float>(boundary.getMaxY());
 	}
 	else if (y < boundary.getMinY()) {
-		over = y - boundary.getMinY();
+		over = y - static_cast<float>(boundary.getMinY());
 	}
 	return over;
 }
@@ -139,10 +139,10 @@ float PBSPHBoundarySolver::getOverZ(const float z)
 {
 	float over = 0;
 	if (z > boundary.getMaxZ()) {
-		over = z - boundary.getMaxZ();
+		over = z - static_cast<float>(boundary.getMaxZ());
 	}
 	else if (z < boundary.getMinZ()) {
-		over = z - boundary.getMinZ();
+		over = z - static_cast<float>(boundary.getMinZ());
 	}
 	return over;
 }
